rpc: backport: add openrpc.discover method support
This changeset is based on the concept originally
implemented at https://github.com/multi-geth/multi-geth/pull/76/files.

It has been adopted to use the present iteration of the rpc/ package,
which varies now significantly from the master branch draft, and
is intended to modify this present rpc/ draft as little as possible,
with the expectation that the master branch iteration of the package
will be adopted and this feature implementation will be made irrelevant
since the change has already been implemented in a later draft
which is not yet considered released.